+++ bootique-cayenne-ehcache/src/main/java/io/bootique/cayenne/ehcache/CayenneEhCacheModuleProvider.java
@@ -1,0 +1,12 @@
+package io.bootique.cayenne.ehcache;
+
+import com.google.inject.Module;
+import io.bootique.BQModuleProvider;
+
+public class CayenneEhCacheModuleProvider implements BQModuleProvider {
+
+    @Override
+    public Module module() {
+        return new CayenneEhCacheModule();
+    }
+}