+++ bootique-cayenne-ehcache/src/test/java/io/bootique/cayenne/ehcache/persistent/Table1.java
@@ -1,0 +1,9 @@
+package io.bootique.cayenne.ehcache.persistent;
+
+import io.bootique.cayenne.ehcache.persistent.auto._Table1;
+
+public class Table1 extends _Table1 {
+
+    private static final long serialVersionUID = 1L; 
+
+}