+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 * #27 Unit tests: add a builder of Table objects based on DbEntity
 * #28 Extend TestDataManager for Cayenne
 * #30 Add a way to register a listener of schema creation event
+* #31 New module: bootique-cayenne-ehcache
 
 ## 0.17